requirements.txt
Django==1.9.2
djangorestframework==3.3.2
enum34==1.1.2
Markdown==2.6.5
psycopg2==2.6.1
